Collision
The code can detect collision now, but it doesn't have the feature yet of perfect collision where if it goes into the wall, it'll place it right next to the wall. I forgot about that. Also if you move right but you're not perfect in place, it should move you into place. Like if you're scrolling right, it should move you up or down until your player fits into that corridor. I attempted on the logic but holy crap it's hard. Time to watch the office.